Refactor LED to split event and display type
